[Feature #155965773] Move GitHubAdaptor into adaptor package
